e8 - customer - add cart skeleton

+++ src/main/java/com/store/product/model/response/DiscountResponse.java
@@ -5,6 +5,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Data
 @Getter
 @AllArgsConstructor
@@ -12,5 +14,5 @@ public class DiscountResponse {
     private Long id;
     private String type;
     private int quantityRequired;
-    private double discountPercentage;
+    private BigDecimal discountPercentage;
 }
